upgrades to angular 19
