+++ package.json
@@ -7,30 +7,28 @@
     "build": "ng build --configuration=production",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "dev:ssr": "ng run lofi-code:serve-ssr",
-    "serve:ssr": "node dist/lofi-code/server/main.js",
-    "build:ssr": "ng build --configuration serverless && ng run lofi-code:server",
     "deploy": "npm run build:ssr && serverless deploy",
     "tag": "npm version patch -m 'Deploy version %s'",
-    "encode-env-prod": "base64 -w 0 src/environments/environment.prod.ts"
+    "encode-env-prod": "base64 -w 0 src/environments/environment.prod.ts",
+    "serve:ssr:lofi-code": "node dist/lofi-code/server/server.mjs"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^18.2.13",
-    "@angular/cdk": "^18.2.14",
-    "@angular/common": "^18.2.13",
-    "@angular/compiler": "^18.2.13",
-    "@angular/core": "^18.2.13",
-    "@angular/fire": "^18.0.1",
-    "@angular/forms": "^18.2.13",
-    "@angular/platform-browser": "^18.2.13",
-    "@angular/platform-browser-dynamic": "^18.2.13",
-    "@angular/platform-server": "^18.2.13",
-    "@angular/router": "^18.2.13",
-    "@angular/ssr": "^18.2.15",
-    "@angular/youtube-player": "^18.2.14",
+    "@angular/animations": "^19.2.3",
+    "@angular/cdk": "^19.2.6",
+    "@angular/common": "^19.2.3",
+    "@angular/compiler": "^19.2.3",
+    "@angular/core": "^19.2.3",
+    "@angular/fire": "^19.0.0",
+    "@angular/forms": "^19.2.3",
+    "@angular/platform-browser": "^19.2.3",
+    "@angular/platform-browser-dynamic": "^19.2.3",
+    "@angular/platform-server": "^19.2.3",
+    "@angular/router": "^19.2.3",
+    "@angular/ssr": "^19.2.4",
+    "@angular/youtube-player": "^19.2.6",
     "@codegenie/serverless-express": "^4.14.0",
-    "@fortawesome/angular-fontawesome": "^0.15.0",
+    "@fortawesome/angular-fontawesome": "^1.0.0",
     "@fortawesome/fontawesome-free": "^6.5.1",
     "@fortawesome/fontawesome-svg-core": "^6.5.1",
     "@fortawesome/free-brands-svg-icons": "^6.5.1",
@@ -38,7 +36,7 @@
     "@fortawesome/free-solid-svg-icons": "^6.5.1",
     "aws-serverless-express": "^3.4.0",
     "axios": "^1.6.8",
-    "bulma": "^0.9.4",
+    "bulma": "^1.0.3",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.18.2",
@@ -51,12 +49,12 @@
     "ngx-timeago": "^3.0.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "^0.14.10"
+    "zone.js": "^0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.2.15",
-    "@angular/cli": "^18.2.15",
-    "@angular/compiler-cli": "^18.2.13",
+    "@angular-devkit/build-angular": "^19.2.4",
+    "@angular/cli": "^19.2.4",
+    "@angular/compiler-cli": "^19.2.3",
     "@types/express": "^4.17.17",
     "@types/highlight.js": "^10.1.0",
     "@types/jasmine": "~4.3.0",
@@ -74,6 +72,6 @@
     "serverless-associate-waf": "^1.2.1",
     "serverless-domain-manager": "^7.3.8",
     "serverless-offline": "^13.3.3",
-    "typescript": "^5.5.4"
+    "typescript": "^5.8.2"
   }
 }
