+++ src/app/app.component.html
@@ -10,7 +10,7 @@
     </nav>
   </div>
 }
-<section class="has-background-light">
+<section>
   <router-outlet></router-outlet>
 </section>
 <footer class="footer">
