+++ src/app/app.component.ts
@@ -5,11 +5,10 @@ import { Breadcrumb, BreadcrumbService } from './services/breadcrumb.service';
 import { NavbarComponent } from './navbar/navbar.component';
 
 @Component({
-  selector: 'app-root',
-  standalone: true,
-  imports: [CommonModule, RouterOutlet, RouterModule, NavbarComponent],
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+    selector: 'app-root',
+    imports: [CommonModule, RouterOutlet, RouterModule, NavbarComponent],
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.scss']
 })
 export class AppComponent implements OnInit {
   title = 'lofi code';
