+++ src/app/home/home.component.html
@@ -6,15 +6,15 @@
     </p>
   </div>
 </section>
-<section class="section is-small">
-  <div class="container is-max-desktop">
+<section class="section">
+  <div class="container">
     <h3 class="title is-3 mb-5">Browse by Tag</h3>
-    <div class="has-background-white p-3">
+    <div class="p-3">
       <app-tags size="is-large"></app-tags>
     </div>
   </div>
 </section>
-<section class="section is-small">
+<section class="section">
   <div class="container">
     <h3 class="title is-3 mb-5">Recent Posts</h3>
     <div class="columns is-multiline is-flex" style="align-items: stretch;">
