+++ src/app/post-preview/post-preview.component.html
@@ -1,6 +1,6 @@
-<a
+<div
   *ngIf="post"
-  class="box full-height"
+  class="box full-height has-background is-shadowless is-clickable"
   [routerLink]="['/posts', post.slug]"
 >
   <div
@@ -17,7 +17,7 @@
   <div class="card-content">
     <div class="media">
       <div class="media-content">
-        <h5 class="title is-size-5">{{post.name}}</h5>
+        <h5 class="title is-size-5 mt-5">{{post.name}}</h5>
         <h6 class="subtitle is-size-6">{{post.publish_date.seconds * 1000 | date: 'mediumDate'}}</h6>
       </div>
     </div>
@@ -26,4 +26,4 @@
       <app-tags [tags]="post.tags"></app-tags>
     </div>
   </div>
-</a>
+</div>
