+++ src/app/post-preview/post-preview.component.scss
@@ -16,12 +16,10 @@
 }
 
 a .card-content {
-    color: initial;
     text-decoration: none;
 }
 
 a .card-content:hover {
-    color: initial;
     text-decoration: none;
 }
