+++ src/app/posts/posts.component.html
@@ -1,4 +1,4 @@
-<div class="has-background-white p-3">
+<div class="p-3">
   <app-tags ngSkipHydration />
   <hr>
   <section class="container mt-5">
