+++ src/app/privacy-policy/privacy-policy.component.ts
@@ -2,11 +2,10 @@ import { Component } from '@angular/core';
 
 
 @Component({
-  selector: 'app-privacy-policy',
-  standalone: true,
-  imports: [],
-  templateUrl: './privacy-policy.component.html',
-  styleUrls: ['./privacy-policy.component.scss']
+    selector: 'app-privacy-policy',
+    imports: [],
+    templateUrl: './privacy-policy.component.html',
+    styleUrls: ['./privacy-policy.component.scss']
 })
 export class PrivacyPolicyComponent {
