+++ src/app/tos/tos.component.ts
@@ -2,11 +2,10 @@ import { Component } from '@angular/core';
 
 
 @Component({
-  selector: 'app-tos',
-  standalone: true,
-  imports: [],
-  templateUrl: './tos.component.html',
-  styleUrls: ['./tos.component.scss']
+    selector: 'app-tos',
+    imports: [],
+    templateUrl: './tos.component.html',
+    styleUrls: ['./tos.component.scss']
 })
 export class TosComponent {
