--- src/custom-variables.scss
@@ -1,3 +1,0 @@
-// custom-variables.scss
-$primary: #F89B29;
-// Define other custom variables if needed
