+++ src/styles.scss
@@ -1,12 +1,11 @@
 // styles.scss
-
-@import 'custom-variables';
-@import 'node_modules/bulma/sass/utilities/initial-variables';
-@import 'node_modules/bulma/sass/utilities/functions';
-@import 'node_modules/bulma/sass/utilities/derived-variables';
-@import 'node_modules/bulma/bulma.sass';
-@import 'node_modules/ngx-sharebuttons/themes/default';
-@import 'node_modules/highlight.js/scss/androidstudio';
+$primary: #F89B29;
+@use 'bulma/sass' with (
+  $primary: $primary,
+  $family-primary: '"IBM Plex Sans", sans-serif',
+);
+@use 'ngx-sharebuttons/themes/default';
+@use 'highlight.js/scss/androidstudio';
 
 @font-face {
   font-family: 'IBM Plex Sans';
@@ -16,5 +15,5 @@
 }
 
 body {
-    font-family: 'IBM Plex Sans', sans-serif;
+  font-family: 'IBM Plex Sans', sans-serif;
 }
