--- tsconfig.server.json
@@ -1,14 +1,0 @@
-/* To learn more about this file see: https://angular.io/config/tsconfig. */
-{
-  "extends": "./tsconfig.app.json",
-  "compilerOptions": {
-    "outDir": "./out-tsc/server",
-    "types": [
-      "node"
-    ]
-  },
-  "files": [
-    "src/main.server.ts",
-    "server.ts"
-  ]
-}
